added check to ensre not overwriting extended functions, and added custom erros

+++ src/server/models/mongodbBasicOperations.js
@@ -4,10 +4,16 @@ var assert= require("assert");
 //var nodeConstants = require('../conf/constants.json').nodeConstants;
 var dbConstants = require("../lib/conf/constants.json").dbConstants;
 var nodeConstants = require('../lib/conf/constants.json').nodeConstants;
+var utils = require("../lib/utils");
 
+//errors
+
+var MissingParameterError = require("../lib/errors/MissingParameterError");
+var InvalidParameterError = require("../lib/errors/InvalidParameterError");
 
 function mongodbBasicOperations(db,logger){
-	assert.notStrictEqual(db,undefined);
+	if (db === undefined)
+		throw new MissingParameterError("database object is required");
 	this.db = db;
 	if (!logger) logger = require('../lib/logger')(nodeConstants.LOG_DIR)//require('./logger')(nodeConstants.LOG_DIR)//
 	this.logger = logger; //Accepts a logger instance
@@ -17,24 +23,25 @@ mongodbBasicOperations.prototype.find= function(collectionName, query, fields, o
 	var args = arguments;
 	var _this = this;
 	// validate input
-	assert(Object.prototype.toString.call(collectionName) == "[object String]",
-		"Invalid collection");
-	assert(Object.prototype.toString.call(query) == "[object Object]",
-		"Invalid query");
+	if (!collectionName || !query)
+		throw new MissingParameterError("Required parameter is missing");
+	if (!utils.isString(collectionName))
+		throw new InvalidParameterError("Invalid collection name, must be a valid string");
+	if (!utils.isObject(query))
+		throw new InvalidParameterError("Query must be an object");
 
 	if (options){
 		if(options == {})
 			options = undefined;
-
-		else if (options)
-			assert(Object.prototype.toString.call(options) == "[object Object]",
-				"Invalid Options");
+		else if (!utils.isObject(options))
+			throw new InvalidParameterError("Options must be an object")
+			
 	}
 	var promise= new Promise(function(resolve, reject) {
 		_this.db.getDB().collection(collectionName)
 		.find(query, fields, options)
 		.toArray(function(err, doc) {
-			if (doc === null ) doc == [];
+			if (doc === null ) doc = [];
 			if (err) {
 				_this.logger('err',err,{action:'find',arguments:args});
 				reject(err);
@@ -47,15 +54,17 @@ mongodbBasicOperations.prototype.find= function(collectionName, query, fields, o
 };
 
 
-	/* find and remove a patient where options are the query to submit
- 	* returns a promise */
+/* find and remove a patient where options are the query to submit
+* returns a promise */
 mongodbBasicOperations.prototype.removeDocument = function(collectionName,options,user){
 	var args = arguments;
 	var _this = this;
 
 	// validate input
-	assert(Object.prototype.toString.call(options) == "[object Object]",
-		"Invalid Options");
+	if (!collectionName || !options)
+			throw new MissingParameterError("Required parameter is missing");
+	if ( !utils.isObject(options))
+		throw new InvalidParameterError("Required paramter: Options in removeDocument must be an object");
 
 	this.logger('info','removing document from collection', {'collection':collectionName,query:options});
 	var promise = new Promise(function(resolve,reject){
@@ -76,8 +85,14 @@ mongodbBasicOperations.prototype.removeDocument = function(collectionName,option
 mongodbBasicOperations.prototype.aggregate = function(collectionName,aggArray,user){
 		var args = arguments;
 		var _this = this;
-		assert(Object.prototype.toString.call(aggArray) == "[object Array]",
-			"Invalid Options, aggregate requires an array");
+		
+		if (!collectionName || !aggArray)
+			throw new MissingParameterError("Required parameter is missing");
+		if (!utils.isString(collectionName))
+			throw new InvalidParameterError("collectionName must be a valid string");
+		if (!utils.isArray(aggArray))
+			throw new InvalidParameterError("Invalid options, aggregation array must be an array");
+
 
 		var promise = new Promise(function(resolve,reject){
 			var collection = _this.db.getDB().collection(collectionName);
@@ -98,10 +113,12 @@ mongodbBasicOperations.prototype.insert= function(collectionName, doc, user) {
 		var args = arguments;
 		var _this = this;
 		// validate input
-		assert(Object.prototype.toString.call(collectionName) == "[object String]",
-			"Invalid collection");
-		assert(Object.prototype.toString.call(doc) == "[object Object]",
-			"Invalid document");
+		if (!collectionName || !doc)
+			throw new MissingParameterError("Required parameter is missing");
+		if (!utils.isString(collectionName))
+			throw new InvalidParameterError("collectionName must be a valid string");
+		if (!utils.isObject(doc))
+			throw new InvalidParameterError("doc must be an object containing new document information");
 
 		var promise= new Promise(function(resolve, reject) {
 			_this.db.getDB().collection(collectionName).insert(doc, {}, function(err, result) {
@@ -123,12 +140,19 @@ mongodbBasicOperations.prototype.insert= function(collectionName, doc, user) {
  * Returns a promise. */
 mongodbBasicOperations.prototype.insertMany = function(options,user){
 	var _this = this;
-	assert(Object.prototype.toString.call(options) == "[object Object]",
-		"Invalid Options");
+	if (!options)
+		throw new MissingParameterError("Required parameter is missing");
+	if (!options.collectionName || options.documents)
+		throw new MissingParameterError("No Collection Name Provided");
+	if (!utils.isObject(options))
+		throw new InvalidParameterError("Options must be an Object");
+	if (!utils.isString(options.collectionName))
+		throw new InvalidParameterError("collectionName must be a valid string");
+	if (!utils.isArray(options.documents))
+		throw new InvalidParameterError("documents must be in an array");
+
 	var promise = new Promise(function(resolve,reject){
-		if(!options.collectionName){
-			reject(new ReferenceError("No Collection Name Provided"));
-		}
+		
 		_this.db.getDB().collection(options.collectionName,function(err,collection){
 			var bulk = collection.initializeOrderedBulkOp();
 			for (var i = 0; i < options.documents.length; i++){
@@ -157,15 +181,18 @@ mongodbBasicOperations.prototype.update= function(collectionName, query, doc, op
 	var _this = this;
 
 	// validate input
-	assert(Object.prototype.toString.call(collectionName) == "[object String]",
-		"Invalid collection");
-	assert(Object.prototype.toString.call(query) == "[object Object]",
-		"Invalid query");
-	assert(Object.prototype.toString.call(doc) == "[object Object]",
-		"Invalid update document");
+	if (! collectionName || ! query || !doc)
+		throw new MissingParameterError("Required parameters are missing");
+	if (!utils.isString(collectionName))
+		throw new InvalidParameterError("collectionName must be a valid string");
+	if (!utils.isObject(query))
+		throw new InvalidParameterError("Query must be an object");
+	if (!utils.isObject(doc))
+		throw new InvalidParameterError("Document to update must be an object");
 	if (options)
-		assert(Object.prototype.toString.call(options) == "[object Object]",
-		"Invalid update options");
+		if (!utils.isObject(options))
+			throw new InvalidParameterError("Options must be an object");
+
 	var promise= new Promise(function(resolve, reject) {
 		_this.db.getDB().collection(collectionName).update(query, doc, options, function(err, resultDoc) {
 			if (err) {
@@ -187,10 +214,12 @@ mongodbBasicOperations.prototype.createIndex= function(collectionName, spec, opt
 	var args = arguments;
 	var _this = this;
 	// validate input
-	assert(Object.prototype.toString.call(collectionName) == "[object String]",
-		"Invalid collection");
-	assert(Object.prototype.toString.call(spec) == "[object Object]",
-		"Invalid spec");
+	if (! collectionName || ! spec )
+		throw new MissingParameterError("Required parameters are missing");
+	if (!utils.isString(collectionName))
+		throw new InvalidParameterError("collectionName must be a valid string");
+	if (!utils.isObject(spec))
+		throw new InvalidParameterError("Invalid spec");
 
 	var promise= new Promise(function(resolve, reject) {
 		_this.db.getDB().collection(collectionName).createIndex(spec, options, function(err, result) {
@@ -210,11 +239,14 @@ mongodbBasicOperations.prototype.createIndex= function(collectionName, spec, opt
 mongodbBasicOperations.prototype.findOne= function(collectionName,query,user) {
 	var args = arguments;
 	var _this = this;
+	if (! collectionName || ! query)
+		throw new MissingParameterError("Required parameters are missing");
 	// validate input
-	assert(Object.prototype.toString.call(collectionName) == "[object String]",
-		"Invalid collection");
-	assert(Object.prototype.toString.call(query) == "[object Object]",
-		"Invalid query");
+	if (!utils.isString(collectionName))
+		throw new InvalidParameterError("collectionName must be a valid String");
+	if (!utils.isObject(query))
+		throw new InvalidParameterError("Query must be an object");
+
 	var promise= new Promise(function(resolve, reject) {
 		_this.db.getDB().collection(collectionName).findOne(query,function(err, doc) {
 			if (err) {
@@ -237,6 +269,12 @@ mongodbBasicOperations.prototype.findOne= function(collectionName,query,user) {
  * Returns a promise. */
 mongodbBasicOperations.prototype.count= function(collectionName, query) {
 	var _this = this;
+	if (!collectionName || !query)
+		throw new MissingParameterError("Required parameters are missing");
+	if (!utils.isString(collectionName))
+		throw new InvalidParameterError("collectionName must be a valid string");
+	if (!utils.isObject(query))
+		throw new InvalidParameterError("Query must be an object");
 	var promise= new Promise(function(resolve, reject) {
 		_this.db.getDB().collection(collectionName).count(query, function(err, count){
 			if (err) {
@@ -253,6 +291,9 @@ mongodbBasicOperations.prototype.count= function(collectionName, query) {
  * returns a promise */
 mongodbBasicOperations.prototype.createCollection = function(name,user){
 	var _this = this;
+	if (!name)
+		throw new MissingParameterError("A collection name is required");
+	if (!utils.isString(name))
 	var promise = new Promise(function(resolve,reject){
 		if (name){
 			_this.db.getDB().createCollection(name,{strict:true},function(err,collection){
@@ -278,8 +319,10 @@ mongodbBasicOperations.prototype.createCollection = function(name,user){
  * returns a promise */
 mongodbBasicOperations.prototype.dropCollection = function(collectionName,user){
 	var _this = this;
-	assert(Object.prototype.toString.call(collectionName) == "[object String]",
-		"Invalid Options");
+	if (!collectionName)
+		throw new MissingParameterError("CollectionName required");
+	if (!utils.isString(collectionName))
+		throw new InvalidParameterError("collectionName must be a valid string");
 
 	var promise = new Promise(function(resolve,reject){
 		_this.db.getDB().dropCollection(collectionName, function(err,done){
@@ -297,10 +340,13 @@ mongodbBasicOperations.prototype.dropCollection = function(collectionName,user){
 
 /* Check within the specified database to determine whether or not an item exists*/
 mongodbBasicOperations.prototype.checkInDatabase = function(collection,field,value,user){
-	assert(Object.prototype.toString.call(collection) == "[object String]",
-		"Invalid collection");
-	assert(Object.prototype.toString.call(field) == "[object String]",
-		"Invalid collection");
+
+	if (!collection || ! field || ! value)
+		throw new MissingParameterError("Required paramter is missing");
+	if (!utils.isString(collection))
+		throw new InvalidParameterError("collection must be a valid string");
+	if (!utils.isString(field))
+		throw new InvalidParameterError("field paramter must be a valid string");
 	var query = {};
 	query[field] = value;
 	return this.findOne(collection,query,user).then(function(result){
@@ -314,10 +360,13 @@ mongodbBasicOperations.prototype.checkInDatabase = function(collection,field,val
 
 //get the owner of a specified field
 mongodbBasicOperations.prototype.getOwner = function(collection,field){
-	assert(Object.prototype.toString.call(collection) == "[object String]",
-		"Invalid collection");
-	assert(Object.prototype.toString.call(field) == "[object String]",
-		"Invalid collection");
+	if (!collection || !field)
+		throw new MissingParameterError("Required parameter missing");
+	if (!utils.isString(collection))
+		throw new InvalidParameterError("collection must be a valid string");
+	if (!utils.isString(field))
+		throw new InvalidParameterError("field must be a string");
+
 	var query = {};
 	if (collection === dbConstants.PROJECTS.COLLECTION){
 		query[dbConstants.PROJECTS.ID_FIELD]=field;
